[Keymap] keebio quefrency Rev3 add German programmer map
This keymap is optimized for german layout and being a lefty programmer.
The space key is on the left side of this split keyboard. the other side is the Mod button for the first layer.

The typical often used bracket pairs for programming are located in the first layer on the left side of the
keyboard to make typing these much faster instead of always bending your fingers to the AltGr key.

Backspace and Delete are too on the left side of the board located on the first layer.

The arrow keys are also additionally mapped at the first layer on the H, J, K, L keys for smooth integration
in a Vim-User workflow.

Signed-Off-by: Marc Schätz <marc@schaetz.info>